ML-WS/requirements.txt
Flask
torch
torchvision
Pillow
numpy
timm
Werkzeug
